% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_pagespeed.R
\name{download_pagespeed}
\alias{download_pagespeed}
\title{Title}
\usage{
download_pagespeed(url, strategy = NULL, output_type = "simple",
  interval = 0.5, keep_tmp = FALSE,
  key = Sys.getenv("PAGESPEED_API_KEY"), filter_third_party = NULL,
  locale = NULL, rule = NULL, screenshot = NULL, snapshots = NULL,
  utm_campaign = NULL, utm_source = NULL, fields = NULL)
}
\arguments{
\item{fields}{}
}
\description{
Title
}
