% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lighthouse_general_info.R
\name{extract_lighthouse_general_info}
\alias{extract_lighthouse_general_info}
\title{Extract general information for all URLs in Lighthouse report object.}
\usage{
extract_lighthouse_general_info(x)
}
\arguments{
\item{x}{downloaded LH report object}
}
\value{
data frames with the scores
}
\description{
Extract general information for all URLs in Lighthouse report object.
}
\examples{
\dontrun{
# tbd
}
}
