% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lighthouse_scores.R
\name{extract_lighthouse_scores}
\alias{extract_lighthouse_scores}
\title{Extract category scores for all URLs in Lighthouse report object.}
\usage{
extract_lighthouse_scores(x)
}
\arguments{
\item{x}{nested list object returned by \code{download_lighthouse()} function}
}
\value{
data frames with the scores
}
\description{
Extract category scores for all URLs in Lighthouse report object.
}
\examples{
\dontrun{
# download Lighthouse report as a global enviroment object
lh_object <- download_lighthouse(
url        = c("https://www.w3.org/",
               "https://www.google.com"),
categories = c("performance",
               "accessibility",
               "best-practices",
               "pwa",
               "seo"))

# download Lighthouse report as a reference to file
lh_ref <- download_lighthouse(
  url            = c("https://www.w3.org/",
                     "https://www.google.com"),
  categories     = c("performance",
                     "accessibility",
                     "best-practices",
                     "pwa",
                     "seo"),
  as_reference   = TRUE,
  reference_path = "ref_path.llo")

# extraction
scores_obj  <- lh_score_extraction(lh_object)
scores_ref <- lh_score_extraction(lh_ref)

# let's see the objects' structures!

str(scores_obj)
'data.frame':	2 obs. of  7 variables:
$ url           : chr  "https://www.w3.org/" "https://www.google.com/"
$ device        : chr  "desktop" "desktop"
$ performance   : num  1 1
$ accessibility : num  0.78 0.87
$ best.practices: num  0.77 0.85
$ seo           : num  1 0.8
$ pwa           : num  0.54 0.46

> str(scores_ref)
'data.frame':	2 obs. of  7 variables:
$ url           : chr  "https://www.w3.org/" "https://www.google.com/"
$ device        : chr  "desktop" "desktop"
$ performance   : num  1 1
$ accessibility : num  0.78 0.89
$ best.practices: num  0.77 0.92
$ seo           : num  1 0.8
$ pwa           : num  0.54 0.46#'
}
}
