% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_lh_basic_extract.R
\name{fun_lh_basic_extract}
\alias{fun_lh_basic_extract}
\title{Extraction of basic details from Lighthouse reports}
\usage{
fun_lh_basic_extract(audits, report_cat_df)
}
\arguments{
\item{audits}{audits object}

\item{report_cat_df}{reports category object}
}
\value{
data frame with basic Lighthouse columns
}
\description{
Extraction of basic details from Lighthouse reports
}
