% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_lh_cat_id.R
\name{fun_lh_cat_id}
\alias{fun_lh_cat_id}
\title{Helper function for extracting audit categories}
\usage{
fun_lh_cat_id(id, parsed)
}
\arguments{
\item{id}{id of the category}

\item{parsed}{parsed LH output object}
}
\value{
data frame
}
\description{
Helper function for extracting audit categories
}
\examples{
\dontrun{
for(i in 1:length(categories)){
  res <- fun_lh_extract_cat_id(categories[i], parsed)
  report_cat_df <- rbind(report_cat_df, res)
}

str(report_cat_df)
'data.frame':	198 obs. of  2 variables:
$ category   : chr  "performance" "performance" "performance" "performance"
    ...
$ report_name: chr  "first_contentful_paint" "first_meaningful_paint"
    "speed_index" "interactive" ...
}
}
