% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_lh_enhanced_extract.R
\name{fun_lh_enhanced_extract}
\alias{fun_lh_enhanced_extract}
\title{Extraction of enhanced details from Lighthouse reports}
\usage{
fun_lh_enhanced_extract(audits, categories, report_cat_df)
}
\arguments{
\item{audits}{audits object}

\item{categories}{Lighthouse categories to be run}

\item{report_cat_df}{reports category object}
}
\value{
data frame with enhanced columns
}
\description{
Extraction of enhanced details from Lighthouse reports
}
