% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_lh_extract_general_info.R
\name{fun_lh_extract_general_info}
\alias{fun_lh_extract_general_info}
\title{Extract the general information from single URL Lighthouse report}
\usage{
fun_lh_extract_general_info(x)
}
\arguments{
\item{x}{object of single report}
}
\value{
data frame
}
\description{
Extract the general information from single URL Lighthouse report
}
\examples{
\dontrun{
res <- fun_lh_extract_general_info(x = obj)

}
}
