% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_lh_extract_scores.R
\name{fun_lh_extract_scores}
\alias{fun_lh_extract_scores}
\title{Extract the category scores from single URL Lighthouse report}
\usage{
fun_lh_extract_scores(x)
}
\arguments{
\item{x}{LH report object for a single URL}
}
\value{
data frame
}
\description{
Extract the category scores from single URL Lighthouse report
}
\examples{
\dontrun{
# download Lighthouse report as a global enviroment object
lh_object <- download_lighthouse(
url        = c("https://www.w3.org/"),
categories = c("performance",
               "accessibility",
               "best-practices",
               "pwa",
               "seo"))

# download Lighthouse report as a reference to file
lh_ref <- download_lighthouse(
  url            = c("https://www.w3.org/"),
  categories     = c("performance",
                     "accessibility",
                     "best-practices",
                     "pwa",
                     "seo"),
  as_reference   = TRUE,
  reference_path = "ref_path.llo")

# extraction
scores_obj <- fun_lh_extract_scores(lh_object)
scores_ref <- fun_lh_extract_scores(lh_ref)
}
}
