% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_lh_exist.R
\name{fun_lhex}
\alias{fun_lhex}
\title{Checking existence in LH nested list}
\usage{
fun_lhex(report, elem = NULL, x = "results")
}
\arguments{
\item{report}{searched report}

\item{elem}{searched element}

\item{x}{list of all LH reports}
}
\value{
TRUE if exists, FALSE if not
}
\description{
Checking existence in LH nested list
}
