% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_ps_basic_sort.R
\name{fun_ps_basic_sort}
\alias{fun_ps_basic_sort}
\title{Sorting data frame's columns alphabetically to effectively bind them later}
\usage{
fun_ps_basic_sort(x)
}
\arguments{
\item{x}{data frane which columns need to be sorted}
}
\value{
data frame
}
\description{
Sorting data frame's columns alphabetically to effectively bind them later
}
