% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagespeed_raw_v5.R
\name{lh_raw_1}
\alias{lh_raw_1}
\title{Download Pagespeed v4 raport for an URL as a nested list}
\usage{
lh_raw_1(url, key = Sys.getenv("PAGESPEED_API_KEY"), strategy = NULL,
  categories = "performance", interval = 0.5, locale = NULL,
  utm_campaign = NULL, utm_source = NULL)
}
\arguments{
\item{url}{string. The URL to fetch and analyze}

\item{key}{string. Pagespeed API key to authenticate. Defaults to
"PAGESPEED_API_KEY" enviroment variable.}

\item{strategy}{string. The analysis strategy to use. Options: "desktop" or
"mobile". Defaults to "desktop"}

\item{categories}{string. A Lighthouse category to run. Defaults to
"performance". See more in Details section}

\item{interval}{numeric. Number of seconds to wait between multiple queries.
Defaults to 0.5 second.}

\item{locale}{string. The locale used to localize formatted results}

\item{utm_campaign}{string. Campaign name for analytics. Defaults to NULL}

\item{utm_source}{string. Campaign source for analytics. Defaults to NULL}
}
\value{
nested list
}
\description{
This function can check a single URL (character) and parse
   the output into a data frame. This data frame contain all the possible
   information from Pagespeed ver 5.
}
\details{
This function uses version 5 of the API.
   Check function \code{ps_raw_1()} for version 4.
   If you need less information but in form of a data frame,
   use \code{lh_simple_1()}.

The \code{category} parameter regulates which of the tests'
    categories from Lighthouse are to be run. You can select more than
    one.
    Options: "accessibility", "best-practices", "performance", "pwa",
    "seo".
}
\examples{
\dontrun{
single_url_raw_output_5 <- lh_raw_1("https://www.google.com/")
}
}
