% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lh_simple_1_s.R
\name{lh_simple_1}
\alias{lh_simple_1}
\title{Download Pagespeed v4 raport for an URL as a data frame}
\usage{
lh_simple_1(url, key = Sys.getenv("PAGESPEED_API_KEY"),
  strategy = NULL, categories = "performance", interval = 0.5,
  locale = NULL, utm_campaign = NULL, utm_source = NULL)
}
\arguments{
\item{url}{string. The URL to fetch and analyze}

\item{key}{string. Pagespeed API key to authenticate. Defaults to
"PAGESPEED_API_KEY" enviroment variable.}

\item{strategy}{string. The analysis strategy to use. Options: "desktop" or
"mobile". Defaults to "desktop"}

\item{categories}{string. A Lighthouse category/categories to run.
Defaults to "performance". See more in Details section}

\item{interval}{numeric. Number of seconds to wait between multiple queries.
Defaults to 0.5 second.}

\item{locale}{string. The locale used to localize formatted results}

\item{utm_campaign}{string. Campaign name for analytics. Defaults to NULL}

\item{utm_source}{string. Campaign source for analytics. Defaults to NULL}
}
\value{
data frame
}
\description{
Download Pagespeed v4 raport for an URL as a data frame
}
\details{
The \code{category} parameter regulates which of the tests'
    categories from Lighthouse are to be run. You can select more than
    one.
    Options: "accessibility", "best-practices", "performance", "pwa",
    "seo".
}
\examples{
\dontrun{
single_url_simple_output_5 <- lh_simple_1("https://www.google.com/")
}
}
