% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagespeed_raw_list_v5.R
\name{pagespeed_raw_list_v5}
\alias{pagespeed_raw_list_v5}
\title{Download Pagespeed v5 raport (Lighthouse) for multiple URLs as a one nested list}
\usage{
pagespeed_raw_list_v5(url, key = Sys.getenv("PAGESPEED_API_KEY"),
  strategy = NULL, categories = "performance", interval = 0.5,
  keep_tmp = FALSE, enhanced_lighthouse = FALSE, locale = NULL,
  utm_campaign = NULL, utm_source = NULL)
}
\arguments{
\item{url}{vector of character strings. The URLs to fetch and analyze}

\item{key}{string. Pagespeed API key to authenticate. Defaults to
"PAGESPEED_API_KEY" enviroment variable.}

\item{strategy}{string. The analysis strategy to use. Options: "desktop" or
"mobile". Defaults to "desktop"}

\item{categories}{string. A Lighthouse category/categories to run.
Defaults to "performance". See more in Details section}

\item{interval}{numeric. Number of seconds to wait between multiple queries.
Defaults to 0.5 second.}

\item{keep_tmp}{logical. Set to TRUE if you need to keep temporary Rdata file
with parsed response. Defaults to FALSE}

\item{enhanced_lighthouse}{logical. If TRUE, adds even more columns with
Lighthouse data. Defaults to FALSE}

\item{locale}{string. The locale used to localize formatted results}

\item{utm_campaign}{string. Campaign name for analytics. Defaults to NULL}

\item{utm_source}{string. Campaign source for analytics. Defaults to NULL}
}
\value{
unformatted nested list
}
\description{
This function can check multiple URLs (character vector)
   and parse the output into a data frame. This data frame contain all
   the possible information from Pagespeed ver 5.
}
\details{
This function uses legacy version 4 of the API.
   Check function \code{pagespeed_raw_list_v5} for version 5.
   If you need less information but in form of a data frame,
   use \code{pagespeed_simple_lists_v4}.
}
\examples{
\dontrun{
multiple_urls_raw_output <- pagespeed_raw_list_v4("https://www.google.com/")
}
}
