% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagespeed_simple_list.R
\name{pagespeed_simple_list}
\alias{pagespeed_simple_list}
\title{Title}
\usage{
pagespeed_simple_list(url, strategy = NULL, interval = 0.5,
  api_version = 4, keep_tmp = FALSE,
  key = Sys.getenv("PAGESPEED_API_KEY"), filter_third_party = NULL,
  locale = NULL, rule = NULL, screenshot = NULL, snapshots = NULL,
  utm_campaign = NULL, utm_source = NULL, fields = NULL)
}
\arguments{
\item{fields}{}
}
\description{
Title
}
