% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_simple_2_vec.R
\name{ps_simple_2_vec}
\alias{ps_simple_2_vec}
\title{Download Pagespeed v4 raport for multiple URLs as one data frame}
\usage{
ps_simple_2_vec(url, key = Sys.getenv("PAGESPEED_API_KEY"),
  strategy = NULL, interval = 0.5, filter_third_party = NULL,
  locale = NULL, rule = NULL, screenshot = NULL, snapshots = NULL,
  utm_campaign = NULL, utm_source = NULL)
}
\arguments{
\item{url}{vector of character strings. The URLs to fetch and analyze}

\item{key}{string. Pagespeed API key to authenticate. Defaults to
"PAGESPEED_API_KEY" enviroment variable.}

\item{strategy}{string. The analysis strategy to use. Options: "desktop" or
"mobile". Defaults to "desktop"}

\item{interval}{numeric. Number of seconds to wait between multiple queries.
Defaults to 0.5 second.}

\item{filter_third_party}{logical. Indicates if third party resources should
be filtered out before PageSpeed analysis. Defaults to NULL (= FALSE)}

\item{locale}{string. The locale used to localize formatted results}

\item{rule}{string. A PageSpeed rule to run; if none are given, all rules
are run}

\item{screenshot}{logical. Indicates if binary data containing a screenshot
should be included. Defaults to NULL (= FALSE)}

\item{snapshots}{logical. Indicates if binary data containing snapshot images
should be included. Defaults to NULL (= FALSE)}

\item{utm_campaign}{string. Campaign name for analytics. Defaults to NULL}

\item{utm_source}{string. Campaign source for analytics. Defaults to NULL}
}
\value{
formatted data frame
}
\description{
This function can check multiple URLs given in a vector
   and parse them into a one data frame. This data frame doesn't
   contain all the possible information from Pagespeed as response
   cannot be simply parsed into tabular form, but it contains
   most of the metrics, recommendations and error occurences.
}
\details{
This function uses legacy version 4 of the API.
   Check function \code{lh_simple_2_vec()} for version 5.
   If you need all the information but in form of a nested list,
   use \code{ps_raw_2_vec()}.
}
\examples{
\dontrun{
multiple_urls_simple_output <- ps_simple_2_vec("https://www.google.com/")
}
}
