% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v5_placeholder_basic.R
\name{v5_placeholder_basic}
\alias{v5_placeholder_basic}
\title{Create basic Lighthouse placeholder}
\usage{
v5_placeholder_basic()
}
\value{
data frame with NAs
}
\description{
Create basic Lighthouse placeholder
}
