% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v5_placeholder_enhanced.R
\name{v5_placeholder_enhanced}
\alias{v5_placeholder_enhanced}
\title{Create enhanced Lighthouse placeholder}
\usage{
v5_placeholder_enhanced()
}
\value{
data frame with NAs
}
\description{
Create enhanced Lighthouse placeholder
}
