% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_pagespeed.R
\name{download_pagespeed}
\alias{download_pagespeed}
\title{Title}
\usage{
download_pagespeed(url, key = Sys.getenv("PAGESPEED_API_KEY"),
  output_type = "simple", api_version = 4, strategy = "desktop",
  interval = 0.5, keep_tmp = FALSE, filter_third_party = NULL,
  locale = NULL, rule = NULL, screenshot = NULL, snapshots = NULL,
  utm_campaign = NULL, utm_source = NULL)
}
\arguments{
\item{utm_source}{}
}
\description{
Title
}
