% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_lh_enhanced_extract.R
\name{fun_lh_enhanced_extract}
\alias{fun_lh_enhanced_extract}
\title{Extraction of extended details from Lighthouse reports}
\usage{
fun_lh_enhanced_extract(audits, report_cat_df = NULL)
}
\arguments{
\item{audits}{audits object}

\item{report_cat_df}{reports category object, defaults to null}
}
\value{
data frame with basic Lighthouse columns
}
\description{
Extraction of extended details from Lighthouse reports
}
\examples{
\dontrun{
    results_load_exp <- fun_lh_enhanced_extract(parsed$loadingExperience)
    results_audits   <- fun_lh_enhanced_extract(audits, report_cat_df)
}
}
